Changed temperature input format from int to double.

+++ src/main/java/cz/ptw/temperature/controllers/TemperatureController.java
@@ -39,7 +39,7 @@ public class TemperatureController {
     @RequestMapping(value = "/rest/temp/add", method = RequestMethod.GET, params = {"probeId", "temperature"})
     public
     @ResponseBody
-    String addTemperatureInformation(@RequestParam("probeId") String probeId, @RequestParam("temperature") int temperature) {
+    String addTemperatureInformation(@RequestParam("probeId") String probeId, @RequestParam("temperature") double temperature) {
         temperatureManager.addNewTemperatureInformationRecord(new TemperatureInformation(probeId, temperature));
 
         return "OK";
